+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.1.4] - 2023-05-13
 
+- fix travel through rotated multidim blocks
 - update console enable/disable debug command
 
 ## [1.1.3] - 2023-04-16